# liverpool: Finish GPU work before signaling idle.