Finish pending GPU work before signaling the `GpuIdle` interrupt, to make sure the guest sees the results in case of e.g. readbacks.

This also fixes Ratchet & Clank (2016) same as https://github.com/shadps4-emu/shadPS4/pull/3593, but with a simpler solution that allows continuing to use scheduler deferred operations.

Performance testing compared to main is appreciated to see if this significantly impacts performance or not.