@StevenMiller123 2025-09-13
Seems to be quite a painful performance drop in KINGDOM HEARTS Ⅲ (readbacks and readbackLinearImages enabled, DMA is disabled).
| Main (~270 ms per frame) | PR (~290 ms per frame) |
|------|----|
| <img width="1922" height="1112" alt="image" src="https://github.com/user-attachments/assets/167c63f7-6c96-4e70-a931-1a640b332009" /> | <img width="1922" height="1112" alt="image" src="https://github.com/user-attachments/assets/4a71ec53-0245-43b7-a18c-efe1d7ca8a3a" /> |

These were tested with a locally compiled debug build, so that may skew results?
[CUSA12031 Main.log](https://github.com/user-attachments/files/22313919/CUSA12031.Main.log)
[CUSA12031 PR.log](https://github.com/user-attachments/files/22313920/CUSA12031.PR.log)