@squidbus 2025-09-13
Okay thanks, was worth checking into at least. Maybe there's some better sync point we can use in the future.